@@ -1,6 +1,4 @@
-"""
-Code Summarizer - AI-powered code analysis and summarization tool
-"""
+"""Code Summarizer - AI-powered code analysis and summarization tool."""
 
 __version__ = "1.0.0"
 __author__ = "Code Analysis Team"
